fix: avoid converting json and tests
